move android code to hoc
