+++ src/containers/gui.jsx
@@ -35,8 +35,7 @@ import storage from '../lib/storage';
 import vmListenerHOC from '../lib/vm-listener-hoc.jsx';
 import vmManagerHOC from '../lib/vm-manager-hoc.jsx';
 import cloudManagerHOC from '../lib/cloud-manager-hoc.jsx';
-import costumeLibraryContent from '../lib/libraries/costumes.json';
-import spriteLibraryContent from '../lib/libraries/sprites.json';
+import AndroidHookHOC from '../lib/android-hook-hoc.jsx';
 
 import GUIComponent from '../components/gui/gui.jsx';
 import {setIsScratchDesktop} from '../lib/isScratchDesktop.js';
@@ -55,12 +54,6 @@ class GUI extends React.Component {
         this.setReduxTitle(this.props.projectTitle);
         this.props.onStorageInit(storage);
         this.props.onVmInit(this.props.vm);
-        if (this.props.onSetCostumeAdder) {
-            this.props.onSetCostumeAdder(this.handleAddCostume.bind(this));
-        }
-        if (this.props.onSetSpriteAdder) {
-            this.props.onSetSpriteAdder(this.handleAddSprite.bind(this));
-        }
     }
     componentDidUpdate (prevProps) {
         if (this.props.projectId !== prevProps.projectId && this.props.projectId !== null) {
@@ -75,14 +68,6 @@ class GUI extends React.Component {
             this.props.onProjectLoaded();
         }
     }
-    componentWillUnmount () {
-        if (this.props.onSetCostumeAdder) {
-            this.props.onSetCostumeAdder(null);
-        }
-        if (this.props.onSetSpriteAdder) {
-            this.props.onSetSpriteAdder(null);
-        }
-    }
     setReduxTitle (newTitle) {
         if (newTitle === null || typeof newTitle === 'undefined') {
             this.props.onUpdateReduxProjectTitle(
@@ -91,48 +76,6 @@ class GUI extends React.Component {
         } else {
             this.props.onUpdateReduxProjectTitle(newTitle);
         }
-    }
-    handleAddCostume (md5) {
-        let item;
-        for (const costume of costumeLibraryContent) {
-            if (costume.md5 === md5) {
-                item = costume;
-                break;
-            }
-        }
-        if (!item) {
-            console.log(`Item not found! ${md5}`);
-            return;
-        }
-        const split = item.md5.split('.');
-        const type = split.length > 1 ? split[1] : null;
-        const rotationCenterX = type === 'svg' ? item.info[0] : item.info[0] / 2;
-        const rotationCenterY = type === 'svg' ? item.info[1] : item.info[1] / 2;
-        const vmCostume = {
-            name: item.name,
-            md5: item.md5,
-            rotationCenterX,
-            rotationCenterY,
-            bitmapResolution: item.info.length > 2 ? item.info[2] : 1,
-            skinId: null
-        };
-        return this.props.vm.addCostumeFromLibrary(md5, vmCostume);
-    }
-    handleAddSprite (md5) {
-        let item;
-        for (const sprite of spriteLibraryContent) {
-            if (sprite.md5 === md5) {
-                item = sprite;
-                break;
-            }
-        }
-        if (!item) {
-            console.log(`Item not found! ${md5}`);
-            return;
-        }
-        return this.props.vm.addSprite(JSON.stringify(item.json)).then(() => { 
-            this.props.onActivateTab(BLOCKS_TAB_INDEX);
-        });
     }
     render () {
         if (this.props.isError) {
@@ -186,8 +129,6 @@ GUI.propTypes = {
     loadingStateVisible: PropTypes.bool,
     onProjectLoaded: PropTypes.func,
     onSeeCommunity: PropTypes.func,
-    onSetCostumeAdder: PropTypes.func,
-    onSetSpriteAdder: PropTypes.func,
     onStorageInit: PropTypes.func,
     onUpdateProjectId: PropTypes.func,
     onUpdateProjectTitle: PropTypes.func,
@@ -261,6 +202,7 @@ const WrappedGui = compose(
     LocalizationHOC,
     ErrorBoundaryHOC('Top Level App'),
     FontLoaderHOC,
+    AndroidHookHOC,
     QueryParserHOC,
     ProjectFetcherHOC,
     ProjectSaverHOC,
